Flytter hjelpetekst for spørsmål om søker fra en bestemt måned til en egen fil, slik at koden enklere kan gjenbrukes.
